feature [003]: projectRefactoring - Code correction + organization
