+++ petJournal/petJournal/Assets.xcassets/Colors/petCTA.colorset/Contents.json
@@ -11,24 +11,6 @@
         }
       },
       "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "0xF1",
-          "green" : "0x93",
-          "red" : "0x80"
-        }
-      },
-      "idiom" : "universal"
     }
   ],
   "info" : {
