+++ petJournal/petJournal/Commons/Enums/States/ErrorForgot.swift
@@ -1,0 +1,13 @@
+//
+//  ErrorForgot.swift
+//  petJournal
+//
+//  Created by Marcylene Barreto on 13/06/23.
+//
+
+import Foundation
+
+enum ErrorForgot: Error {
+    case domainError
+    case mailInvalid
+}
