--- petJournal/petJournal/Commons/Protocols/Coordinator/Coordinator.swift
@@ -1,14 +1,0 @@
-//
-//  Coordinator.swift
-//  petJournal
-//
-//  Created by Marcylene Barreto on 27/04/23.
-//
-
-import UIKit
-
-protocol Coordinator: AnyObject {
-    var navigationController: UINavigationController? { get set }
-    
-    func start()
-}
